info
implementation of differences between different noise models with histograms
Gaussian noise = it is a signal noise that has a probability density function (pdf) equal to that of the normal distribution.
Salt & Pepper noise = Here only a few pixels are noisy, but they are very noisy. The effect is similar to sprinkling white and black dots—salt and pepper—on the image. salt and pepper noise can arise in transmitting images over noisy digital links.